Upgrade Kotlin, Gradle and dependencies

+++ readium/navigator/r2-navigator/src/main/java/org/readium/r2/navigator/pager/R2EpubPageFragment.kt
@@ -89,7 +89,7 @@ class R2EpubPageFragment : Fragment() {
                 val topDecile = webView.contentHeight - 1.15 * metrics.heightPixels
                 val bottomDecile = (webView.contentHeight - metrics.heightPixels).toDouble()
 
-                when (scrollY) {
+                when (scrollY.toDouble()) {
                     in topDecile..bottomDecile -> {
                         if (!endReached) {
                             endReached = true
